+++ rossum-mcp/CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 ---
 
 ## [Unreleased] - YYYY-MM-DD
+-
+
+
+## [0.3.2] - 2025-12-14
 - Fixed allowed sideloads for `get_annotation` [#63](https://github.com/stancld/rossum-mcp/pull/63)
 - Refactored tools to return class instances directly instead of `dataclasses.asdict()` conversions,
 leveraging FastMCP's automatic serialization [#64](https://github.com/stancld/rossum-mcp/pull/64)